COMP: removing double/int warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=12703
git-svn-id: http://svn.slicer.org/Slicer4/trunk@12703 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/ChangeTracker/CommandLineApplication/ChangeTrackerCommandLine.cxx
@@ -659,7 +659,9 @@ void InitializeThresholds(tgCMDLineStructure &tg, vtkImageData* image, vtkImageD
       }
     mean = mean/cnt;
 
-    for(idx=hist->GetMin()[0];idx<hist->GetMax()[0];idx++){
+    for(idx=(int)(hist->GetMin()[0]);
+        idx<(int)(hist->GetMax()[0]);
+        idx++){
       float d = mean - idx;
       stdev += d*d*hist->GetOutput()->GetScalarComponentAsFloat(idx,0,0,0);
     }
